+++ Semi/src/service/plan/PlanServiceImpl.java
@@ -206,15 +206,15 @@ public class PlanServiceImpl implements PlanService{
 
 	// 카테고리 별 총합
 	@Override
-	public int getAccountAccTpeCost(Plan plan, AccType accType) {
+	public double getAccountAccTpeCost(Plan plan, AccType accType) {
 		return plandao.sumAccTypeCost(plan, accType);
 	}
 
 	// 한 일정의 가계부 비용 총합
 	@Override
-	public int getAccountTotal(EnumMap<AccType, Integer> accEnumMap) {
-		int total = 0;
-		for(int cost : accEnumMap.values()) {
+	public double getAccountTotal(EnumMap<AccType, Double> accEnumMap) {
+		double total = 0;
+		for(double cost : accEnumMap.values()) {
 			total += cost;
 		}
 		return total;
